+++ src/main/kotlin/com/github/dinbtechit/ngxs/action/editor/codeIntellisense/generate/NgxsCreateSelectors.kt
@@ -1,6 +1,5 @@
-package com.github.dinbtechit.ngxs.action.editor.codeIntellisense
+package com.github.dinbtechit.ngxs.action.editor.codeIntellisense.generate
 
-import com.github.dinbtechit.ngxs.action.editor.NgxsActionType
 import com.github.dinbtechit.ngxs.action.editor.NgxsStatePsiFile
 import com.intellij.codeInsight.CodeInsightActionHandler
 import com.intellij.codeInsight.actions.CodeInsightAction
@@ -8,22 +7,22 @@ import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiFile
 
-class NgxsCreateAction : CodeInsightAction() {
+class NgxsCreateSelectors : CodeInsightAction() {
 
     override fun isValidForFile(project: Project, editor: Editor, file: PsiFile): Boolean {
         return  NgxsStatePsiFile(file.virtualFile, project).isCursorWithinStateClass(editor, file)
     }
 
     fun invoke(project: Project, editor: Editor, file: PsiFile) {
-        NgxsStatePsiFile(file.virtualFile, project).createActionMethodLiveTemplates(
-            editor, file, NgxsActionType.WITHOUT_PAYLOAD
+        NgxsStatePsiFile(file.virtualFile, project).createSelectorsMethodLiveTemplates(
+            editor, file
         )
     }
 
     override fun getHandler(): CodeInsightActionHandler {
         return object : CodeInsightActionHandler {
             override fun invoke(project: Project, editor: Editor, file: PsiFile) {
-                this@NgxsCreateAction.invoke(project, editor, file)
+                this@NgxsCreateSelectors.invoke(project, editor, file)
             }
 
             override fun startInWriteAction(): Boolean {
