+++ src/main/kotlin/com/github/dinbtechit/ngxs/action/editor/codeIntellisense/inspection/NgxsInspection.kt
@@ -2,13 +2,13 @@ package com.github.dinbtechit.ngxs.action.editor.codeIntellisense.inspection
 
 import com.github.dinbtechit.ngxs.action.editor.NgxsActionUtil
 import com.github.dinbtechit.ngxs.action.editor.NgxsStatePsiFile
+import com.github.dinbtechit.ngxs.action.editor.codeIntellisense.inspection.quickfix.NgxsActionDeclarationQuickFix
 import com.intellij.codeInspection.LocalInspectionToolSession
 import com.intellij.codeInspection.ProblemHighlightType
 import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.lang.javascript.inspections.JSInspection
 import com.intellij.lang.javascript.psi.JSElementVisitor
 import com.intellij.lang.javascript.psi.ecma6.ES6Decorator
-import com.intellij.openapi.vfs.LocalFileSystem
 import com.intellij.psi.PsiElementVisitor
 
 class NgxsActionDecoratorVisitor(private val holder: ProblemsHolder): JSElementVisitor() {
@@ -28,14 +28,11 @@ class NgxsActionDecoratorVisitor(private val holder: ProblemsHolder): JSElementV
                 if (actionTypePsiElement != null && NgxsActionUtil.findActionDeclaration(actionTypePsiElement) == null) {
                     val stateFileName = element.containingFile.name
                     val computedActionFileName = "${stateFileName.split(".")[0]}.actions.ts"
-                    val actionsFile = LocalFileSystem.getInstance().findFileByPath(
-                        "${element.containingFile.containingDirectory.virtualFile.path}/$computedActionFileName"
-                    )
                     holder.registerProblem(
                         actionTypePsiElement,
-                        "Cannot find Action class '${actionTypePsiElement.text}'",
+                        "NGXS: Cannot find Action class '${actionTypePsiElement.text}'",
                         ProblemHighlightType.ERROR,
-                        NgxsActionDeclarationQuickFix(computedActionFileName, actionsFile, actionTypePsiElement)
+                        NgxsActionDeclarationQuickFix(computedActionFileName, actionTypePsiElement)
                     )
 
                 }
@@ -46,11 +43,9 @@ class NgxsActionDecoratorVisitor(private val holder: ProblemsHolder): JSElementV
 }
 
 class NgxsInspection : JSInspection() {
-    override fun createVisitor(holder: ProblemsHolder, p1: LocalInspectionToolSession): PsiElementVisitor {
+    override fun createVisitor(holder: ProblemsHolder, session: LocalInspectionToolSession): PsiElementVisitor {
         if (!NgxsStatePsiFile.isNgxsStateFile(holder.project, holder.file.virtualFile))
             return PsiElementVisitor.EMPTY_VISITOR
         return NgxsActionDecoratorVisitor(holder)
     }
-
-
 }
